requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.3.0
audioread==3.0.1
bleak==0.22.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
decorator==5.1.1
distro==1.9.0
eval_type_backport==0.2.2
exif==1.6.0
filelock==3.16.1
frame-sdk==1.2.3
frozenlist==1.5.0
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jiter==0.7.0
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
ollama==0.3.3
openai==1.53.0
openai-whisper==20240930
packaging==24.1
pillow==10.4.0
platformdirs==4.3.6
plum-py==0.8.7
pooch==1.8.2
propcache==0.2.1
pyarrow==18.1.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.19.1
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-CoreBluetooth==10.3.1
pyobjc-framework-libdispatch==10.3.1
python-dotenv==1.0.1
regex==2024.9.11
requests==2.32.3
rich==13.9.4
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.2.0
shellingham==1.5.4
simpleaudio==1.0.4
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.5.0
tiktoken==0.8.0
together==1.3.11
torch==2.5.0
tqdm==4.66.6
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.18.3
